requirements.txt
aiocoap==0.4.7
eltako14bus==0.0.45
numpy==1.26.3
pillow==10.2.0
pyserial==3.5
pyserial-asyncio==0.6
PyYAML==6.0.1
StrEnum==0.4.15
termcolor==2.4.0
tkinterhtml==0.7
